move utilities outside of the test directory
