+++ enigmaLoader.js
@@ -1,0 +1,2 @@
+import { Enigma, utils, eeConstants, Task } from "./enigma-js/lib/enigma-js.node";
+export { Enigma, utils, eeConstants, Task };
