+++ scUtils.js
