+++ test/20_execute_fail_wrong_worker.test.js
@@ -3,7 +3,7 @@ const fs = require("fs");
 import os from "os";
 const path = require("path");
 const Web3 = require("web3");
-import { Enigma, utils, eeConstants, Task } from "./enigmaLoader";
+import { Enigma, utils, eeConstants, Task } from "../enigmaLoader";
 import {
   EnigmaContract,
   EnigmaTokenContract,
@@ -11,10 +11,10 @@ import {
   EnigmaTokenContractAddress,
   proxyAddress,
   ethNodeAddr
-} from "./contractLoader";
+} from "../contractLoader";
 import EthCrypto from "eth-crypto";
 import EventEmitter from "eventemitter3";
-const constants = require("./testConstants");
+const constants = require("../testConstants");
 
 function sleep(ms) {
   return new Promise(resolve => setTimeout(resolve, ms));
